using adjusted dRdt for prediction
because of the limited capacity for Wuhan medical service, the number of confirmed cases was stuck at a bottleneck before 1/26, and released on 1/27 (likely), which explains the spike from 80 on 1/26 to 892 on 1/27. We provided a method for smoothing out the spike and backing out possible confirmed cases for the earlier days. New set of parameters is estimated with the adjusted data.